ai_models/animal-ai/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
fastapi
uvicorn
python-multipart
Pillow
opencv-python-headless
numpy
